Add files via upload
The purpose of the project is to learn the fundamentals of front-end web development and DOM manipulation through vanilla JavaScript. It gives an interactive interface to add, view, and delete saved bookmarks directly in the browser using local storage. The app is focused on ease of use, a clean interface, and functionality on a variety of devices.